Feat : separate frequency
