+++ src/components/manage-task-item/components/ManageTaskItem.tsx
@@ -2,7 +2,6 @@
 
 import clsx from 'clsx';
 import FormField from '../../common/formField';
-import { OptionSelector } from '../../common/dropdown/OptionSelector';
 import CalendarSelect from '../../calendar/CalendarSelect';
 import WeeklySelect from './WeeklySelect';
 import TimePicker from './TimePicker';
@@ -10,12 +9,7 @@ import useManageTaskItem from '../useManageTaskItem';
 import { TaskItemProps } from '../type';
 import Button from '@/components/common/Button';
 import { ModalFooter } from '@/components/common/modal';
-import useModalContext from '@/components/common/modal/core/useModalContext';
-import DangerModal from '@/components/danger-modal';
-
-const FREQUENCY_LIST = ['한 번', '매일', '주 반복', '월 반복'];
-
-const DELETE_FREQUENCY_MODAL_ID = 'delete-frequency';
+import Frequency from './Frequency';
 
 export default function ManageTaskItem({
   task,
@@ -32,18 +26,15 @@ export default function ManageTaskItem({
     isCalendarOpen,
     isTimeOpen,
     select,
-    FREQUENCY_MAP,
+    frequencyDefaultValue,
     handleInputChange,
     handleCalendarDateChange,
     handleFrequencyChange,
-    markFrequencyForDelete,
     createOrEditSubmit,
     toggleDay,
     updateTime,
     closeModal,
   } = useManageTaskItem({ task, groupId, taskListId, isDone, createOrEditModalId });
-
-  const { openModal } = useModalContext();
 
   const createOrEdit = task ? '수정하기' : '만들기';
 
@@ -70,7 +61,7 @@ export default function ManageTaskItem({
 
           <div className="flex flex-col gap-2">
             <div className="flex flex-col gap-4">
-              <label className="text-lg-md">시간 날짜 및 시간</label>
+              <label className="text-lg-md">시작 날짜 및 시간</label>
               <div className="flex gap-2">
                 {select.map((item) => (
                   <div
@@ -99,25 +90,11 @@ export default function ManageTaskItem({
 
           <div className="flex flex-col gap-4">
             <label className="text-lg-md">반복 설정</label>
-            {task ? (
-              <Button
-                onClick={() => openModal(DELETE_FREQUENCY_MODAL_ID)}
-                type="button"
-                variant="danger"
-                size="custom"
-                className="h-10 w-40 rounded-xl"
-              >
-                반복 설정 삭제하기
-              </Button>
-            ) : (
-              <OptionSelector
-                options={FREQUENCY_LIST}
-                defaultValue={FREQUENCY_MAP[taskItem.frequencyType]}
-                size="sm"
-                placement="top-12"
-                onSelect={handleFrequencyChange}
-              />
-            )}
+            <Frequency
+              isEdit={!!task}
+              frequencyDefaultValue={frequencyDefaultValue}
+              handleFrequencyChange={handleFrequencyChange}
+            />
           </div>
 
           {isWeekly && (
@@ -143,19 +120,6 @@ export default function ManageTaskItem({
           </ModalFooter>
         </form>
       </div>
-      <DangerModal
-        modalId={DELETE_FREQUENCY_MODAL_ID}
-        heading="반복 일정을 삭제하시겠어요?"
-        description={
-          <span>
-            삭제된 반복 일정은 복구할 수 없고,
-            <br />
-            반복 일정은 다시 설정할 수 없습니다.
-          </span>
-        }
-        confirmButton="삭제하기"
-        onConfirm={markFrequencyForDelete}
-      />
     </div>
   );
 }
